move view components
