+++ src/lib/vocab/view/Board.svelte
