+++ src/lib/vocab/view/Home.svelte
