+++ src/lib/vocab/view/Room.svelte
