+++ src/lib/vocab/view/components.ts
@@ -1,13 +1,13 @@
 import type {SvelteComponent} from 'svelte';
 
 import type {ViewType} from '$lib/vocab/view/view';
-import Home from '$lib/ui/Home.svelte';
-import Room from '$lib/ui/Room.svelte';
-import Board from '$lib/ui/Board.svelte';
-import Forum from '$lib/ui/Forum.svelte';
-import Notes from '$lib/ui/Notes.svelte';
-import Iframe from '$lib/ui/Iframe.svelte';
-import Voice from '$lib/ui/Voice.svelte';
+import Home from '$lib/vocab/view/Home.svelte';
+import Room from '$lib/vocab/view/Room.svelte';
+import Board from '$lib/vocab/view/Board.svelte';
+import Forum from '$lib/vocab/view/Forum.svelte';
+import Notes from '$lib/vocab/view/Notes.svelte';
+import Iframe from '$lib/vocab/view/Iframe.svelte';
+import Voice from '$lib/vocab/view/Voice.svelte';
 
 export const viewComponents: Record<ViewType, typeof SvelteComponent> = {
 	Home,
